refactor(core): finish edges/extract by-language split
Lift the remaining per-language extractors out of extract/mod.rs into
job-focused siblings, completing the decomposition:

- extract/rust.rs       — rust_edges + rust_impl_edges (calls the
                          rust_dispatch entry points via use super::rust_dispatch)
- extract/typescript.rs — typescript_edges
- extract/kotlin.rs     — kotlin_edges
- extract/c_like.rs     — c_like_edges

syntactic_edges dispatches to each via <lang>::<lang>_edges; the entry
points are pub(super). extract/mod.rs is now just the dispatcher
(index_file_edges / edge_candidates / syntactic_edges / collect_edges)
and the shared edge-builders (symbol_edge, file_edge, scope helpers).

Pure code movement, logic unchanged. extract/mod.rs: 819 -> 297 lines
(2619 at the start of this branch). fmt, clippy --all-targets, nextest
-p rag-rat-core (868 tests) green.
