+++ crates/rag-rat-core/src/index/edges/extract/c_like.rs
@@ -1,0 +1,70 @@
+//! C / C++ graph-edge extraction — the `Language::C | Language::Cpp` arm of `syntactic_edges`.
+//! Split out of edges/extract.
+use super::*;
+
+pub(super) fn c_like_edges(
+    text: &str,
+    node: Node<'_>,
+    symbols: &[IndexedSymbol],
+    path: &Path,
+    out: &mut Vec<EdgeCandidate>,
+) {
+    match node.kind() {
+        "preproc_include" => {
+            let include = node_text(node, text)
+                .trim()
+                .trim_start_matches("#include")
+                .trim()
+                .trim_matches(['<', '>', '"'])
+                .to_string();
+            if !include.is_empty() {
+                out.push(file_edge(
+                    path,
+                    node,
+                    text,
+                    include,
+                    EdgeKind::Imports,
+                    EdgeConfidence::NameOnly,
+                ));
+            }
+        },
+        "call_expression" => {
+            let function = node.child_by_field_name("function").unwrap_or(node);
+            let identifiers = identifiers_under(function, text);
+            // Parallel to `identifiers` so the callee `.last()` node matches the string pick (#67).
+            let identifier_nodes = identifier_nodes_under(function);
+            if let Some(name) = identifiers.last().cloned().or_else(|| call_target_name(node, text))
+            {
+                out.push(symbol_edge_with_context(
+                    symbols,
+                    node,
+                    text,
+                    name,
+                    EdgeKind::CallsName,
+                    EdgeConfidence::NameOnly,
+                    EdgeContext {
+                        target_qualified_name: c_like_qualified_name(&identifiers),
+                        receiver_hint: identifiers
+                            .first()
+                            .filter(|_| identifiers.len() > 1)
+                            .cloned(),
+                    },
+                    identifier_nodes.last().copied().map(CalleeRange::of_node),
+                ));
+            }
+        },
+        "type_identifier" | "qualified_identifier" | "namespace_identifier" => {
+            if let Some(name) = last_identifier_text(node, text) {
+                out.push(symbol_edge(
+                    symbols,
+                    node,
+                    name,
+                    EdgeKind::ReferencesType,
+                    EdgeConfidence::NameOnly,
+                    last_identifier_node(node).map(final_segment_node).map(CalleeRange::of_node),
+                ));
+            }
+        },
+        _ => {},
+    }
+}
