+++ crates/rag-rat-core/src/index/edges/extract/mod.rs
@@ -1,7 +1,11 @@
 use super::*;
 
+mod c_like;
+mod kotlin;
 mod python;
+mod rust;
 mod rust_dispatch;
+mod typescript;
 
 pub(crate) fn index_file_edges(
     conn: &Connection,
@@ -128,10 +132,10 @@ pub(crate) fn collect_edges(
         return;
     }
     match language {
-        Language::Rust => rust_edges(text, node, symbols, path, out),
-        Language::TypeScript => typescript_edges(text, node, symbols, path, out),
-        Language::Kotlin => kotlin_edges(text, node, symbols, path, out),
-        Language::C | Language::Cpp => c_like_edges(text, node, symbols, path, out),
+        Language::Rust => rust::rust_edges(text, node, symbols, path, out),
+        Language::TypeScript => typescript::typescript_edges(text, node, symbols, path, out),
+        Language::Kotlin => kotlin::kotlin_edges(text, node, symbols, path, out),
+        Language::C | Language::Cpp => c_like::c_like_edges(text, node, symbols, path, out),
         Language::Python => python::python_edges(text, node, symbols, path, out),
         Language::Markdown => {},
     }
@@ -139,532 +143,6 @@ pub(crate) fn collect_edges(
     let mut cursor = node.walk();
     for child in node.named_children(&mut cursor) {
         collect_edges(language, text, child, symbols, path, out);
-    }
-}
-pub(crate) fn rust_edges(
-    text: &str,
-    node: Node<'_>,
-    symbols: &[IndexedSymbol],
-    path: &Path,
-    out: &mut Vec<EdgeCandidate>,
-) {
-    match node.kind() {
-        "use_declaration" => {
-            let names = identifiers_under(node, text);
-            let is_reexport = node_text(node, text).trim_start().starts_with("pub use ");
-            // Module-aware import scope (#61): a Rust `use` is scoped to its enclosing module body
-            // (or block, for a block-local `use`), not the whole file. Record that scope + the
-            // enclosing module's id on the dedicated import-scope columns so resolution suppresses
-            // a bare reference only inside the `use`'s scope (parent-`mod` `use`s don't
-            // reach a child `mod`). Top-level `use` → whole file, `MOD_FILE_ROOT`.
-            let scope = enclosing_use_scope(node, text);
-            // The crate-aware scope rebuilds its {leaf → root} map by re-parsing an Imports edge's
-            // `evidence` with `imports::parse_use`; `parse_use` returns EVERY leaf, so ONE edge
-            // carrying the FULL (untruncated) use text populates the whole map for this `use`.
-            // Attach the full text to only the FIRST emitted Imports edge — the default
-            // `edge_evidence` truncates at 240 chars and would drop late braced leaves (#97 item 1)
-            // — and let the rest carry standard evidence, so a multi-hundred-KB `use` isn't cloned
-            // into every leaf's edge (#97 item 3).
-            let mut full_use_evidence = Some(use_declaration_evidence(node, text));
-            for name in names {
-                if !is_rust_path_keyword(&name) {
-                    let evidence =
-                        full_use_evidence.take().unwrap_or_else(|| edge_evidence(node, text));
-                    out.push(file_edge_scoped(
-                        path,
-                        node,
-                        name,
-                        Some(evidence),
-                        EdgeKind::Imports,
-                        EdgeConfidence::NameOnly,
-                        Some(scope),
-                    ));
-                }
-            }
-            if is_reexport {
-                for name in identifiers_under(node, text) {
-                    if !is_rust_path_keyword(&name) {
-                        out.push(file_edge(
-                            path,
-                            node,
-                            text,
-                            name,
-                            EdgeKind::Exports,
-                            EdgeConfidence::NameOnly,
-                        ));
-                    }
-                }
-            }
-        },
-        "mod_item" =>
-            if let Some(name) = child_name_text(node, text) {
-                // An INLINE `mod foo { … }` carries its body range + its own id as the import scope
-                // so resolution can rebuild the per-file module interval set (the ref→mod-id
-                // lookup) from edges alone, WITHOUT the tree (#61) — including modules that contain
-                // no `use`. A non-inline `mod foo;` has no body and introduces no scope (NULL).
-                let scope = inline_mod_scope(node);
-                out.push(file_edge_scoped(
-                    path,
-                    node,
-                    name,
-                    Some(edge_evidence(node, text)),
-                    EdgeKind::Imports,
-                    EdgeConfidence::NameOnly,
-                    scope,
-                ));
-            },
-        "call_expression" => {
-            if let Some(name) = call_target_name(node, text) {
-                out.push(symbol_edge_with_context(
-                    symbols,
-                    node,
-                    text,
-                    name,
-                    EdgeKind::CallsName,
-                    EdgeConfidence::NameOnly,
-                    EdgeContext {
-                        target_qualified_name: target_qualified_name(node, text),
-                        receiver_hint: scoped_receiver_name(node, text),
-                    },
-                    call_target_node(node).map(CalleeRange::of_node),
-                ));
-            }
-            // A scoped call receiver is a type reference only when it names a type. By Rust
-            // convention types are PascalCase, while module paths (`std::env::…`) and method
-            // receivers on locals (`p.as_os_str()`) are snake_case — emitting those as
-            // `references_type` produced bogus "types" like `std` and `p`. Gate on an
-            // uppercase-leading receiver so `Foo::bar()` still records a type reference.
-            if let Some(receiver) = scoped_receiver_name(node, text)
-                && receiver.chars().next().is_some_and(char::is_uppercase)
-            {
-                out.push(symbol_edge(
-                    symbols,
-                    node,
-                    receiver,
-                    EdgeKind::ReferencesType,
-                    EdgeConfidence::NameOnly,
-                    // The type is the receiver — the LEADING `::` segment (`Foo` in `Foo::bar()`)
-                    // — so anchor the range on the function path's first
-                    // identifier, not its tail.
-                    node.child_by_field_name("function")
-                        .and_then(first_identifier_node)
-                        .map(CalleeRange::of_node),
-                ));
-            }
-            // #200 dispatch construct fact: a tuple enum-variant construction `Enum::Variant(..)`.
-            // Key off the FULL call path's last two PascalCase `::` segments, so
-            // `crate::m::Msg::Start(..)` still yields `Msg::Start` (the bare receiver would be
-            // `crate`). `Foo::new()` / `T::CONST` are excluded (tail not PascalCase).
-            if let Some(key) = node
-                .child_by_field_name("function")
-                .and_then(|f| rust_dispatch::enum_variant_key(f, text))
-            {
-                out.push(rust_dispatch::dispatch_fact(
-                    symbols,
-                    node,
-                    key,
-                    EdgeKind::DispatchConstruct,
-                    EdgeContext::default(),
-                    None,
-                ));
-            }
-        },
-        "struct_expression" =>
-        // #200 dispatch construct fact: a struct enum-variant construction `Enum::Variant { .. }`.
-        {
-            if let Some(key) = node
-                .child_by_field_name("name")
-                .and_then(|n| rust_dispatch::enum_variant_key(n, text))
-            {
-                out.push(rust_dispatch::dispatch_fact(
-                    symbols,
-                    node,
-                    key,
-                    EdgeKind::DispatchConstruct,
-                    EdgeContext::default(),
-                    None,
-                ));
-            }
-        },
-        "scoped_identifier" =>
-        // #200 dispatch construct fact: a UNIT enum-variant `Enum::Stop` in a VALUE position — a
-        // call argument (`send(Msg::Stop)`), a `let` initializer (`let m = Msg::Stop;`), an
-        // assignment RHS, or a `return`/`break` value. Not a `call_expression`, so the arms above
-        // miss it. The value-position gate keeps an ordinary type/module/use path out;
-        // over-emitting for a non-enum `Foo::Bar` is harmless — synthesis only joins a
-        // variant whose head is a unique in-scope `enum`, so a non-enum head never yields a
-        // `dispatches` edge.
-        {
-            if rust_dispatch::scoped_identifier_in_value_position(node)
-                && let Some(key) = rust_dispatch::enum_variant_key(node, text)
-            {
-                out.push(rust_dispatch::dispatch_fact(
-                    symbols,
-                    node,
-                    key,
-                    EdgeKind::DispatchConstruct,
-                    EdgeContext::default(),
-                    None,
-                ));
-            }
-        },
-        "match_arm" => rust_dispatch::rust_dispatch_handle_facts(text, node, symbols, out),
-        "macro_invocation" =>
-            if let Some(name) = first_identifier_text(node, text) {
-                out.push(symbol_edge_with_context(
-                    symbols,
-                    node,
-                    text,
-                    name,
-                    EdgeKind::UsesMacro,
-                    EdgeConfidence::NameOnly,
-                    EdgeContext::default(),
-                    first_identifier_node(node).map(CalleeRange::of_node),
-                ));
-            },
-        "impl_item" => rust_impl_edges(text, node, symbols, out),
-        "type_identifier" | "scoped_type_identifier" | "generic_type" => {
-            if let Some(name) = last_identifier_text(node, text) {
-                out.push(symbol_edge(
-                    symbols,
-                    node,
-                    name,
-                    EdgeKind::ReferencesType,
-                    EdgeConfidence::NameOnly,
-                    last_identifier_node(node).map(final_segment_node).map(CalleeRange::of_node),
-                ));
-            }
-        },
-        _ => {},
-    }
-}
-pub(crate) fn rust_impl_edges(
-    text: &str,
-    node: Node<'_>,
-    symbols: &[IndexedSymbol],
-    out: &mut Vec<EdgeCandidate>,
-) {
-    let node_text = node_text(node, text);
-    let header = node_text.split('{').next().unwrap_or_default();
-    let type_names = header
-        .split(|ch: char| !ch.is_alphanumeric() && ch != '_')
-        .filter(|part| !part.is_empty())
-        .filter(|part| !matches!(*part, "impl" | "for" | "where"))
-        .map(ToOwned::to_owned)
-        .collect::<Vec<_>>();
-    if node_text.contains(" for ") && type_names.len() >= 2 {
-        let trait_name = type_names.first().cloned().unwrap_or_default();
-        let type_name = type_names.last().cloned().unwrap_or_default();
-        out.push(EdgeCandidate {
-            from_symbol_id: containing_symbol(symbols, node.start_byte()).map(|symbol| symbol.id),
-            from_name: Some(type_name),
-            to_name: trait_name,
-            target_qualified_name: None,
-            evidence: Some(edge_evidence(node, text)),
-            receiver_hint: None,
-            source_span: span_for_node(node),
-            // The trait/type names here come from string-splitting the impl header, not from a
-            // located identifier node, so there is no clean callee range to record (#67).
-            callee_span: None,
-            import_scope: None,
-            edge_kind: EdgeKind::Implements,
-            confidence: EdgeConfidence::NameOnly,
-        });
-    } else if let Some(type_name) = type_names.first() {
-        out.push(symbol_edge(
-            symbols,
-            node,
-            type_name.clone(),
-            EdgeKind::ReferencesType,
-            EdgeConfidence::NameOnly,
-            // `type_name` is string-split from the impl header, not a located node — no range
-            // (#67).
-            None,
-        ));
-    }
-}
-pub(crate) fn typescript_edges(
-    text: &str,
-    node: Node<'_>,
-    symbols: &[IndexedSymbol],
-    path: &Path,
-    out: &mut Vec<EdgeCandidate>,
-) {
-    match node.kind() {
-        "import_statement" =>
-            for name in identifiers_under(node, text) {
-                out.push(file_edge(
-                    path,
-                    node,
-                    text,
-                    name,
-                    EdgeKind::Imports,
-                    EdgeConfidence::NameOnly,
-                ));
-            },
-        "export_statement" =>
-            for name in identifiers_under(node, text) {
-                out.push(file_edge(
-                    path,
-                    node,
-                    text,
-                    name,
-                    EdgeKind::Exports,
-                    EdgeConfidence::NameOnly,
-                ));
-            },
-        "call_expression" | "new_expression" => {
-            let function = node.child_by_field_name("function").unwrap_or(node);
-            let identifiers = identifiers_under(function, text);
-            // Parallel to `identifiers` (same traversal order), so `.last()`/`.first()` pick the
-            // node for the same token the string Vec does (#67).
-            let identifier_nodes = identifier_nodes_under(function);
-            if let Some(name) = identifiers.last().cloned().or_else(|| call_target_name(node, text))
-            {
-                let edge_kind = if node.kind() == "new_expression" {
-                    EdgeKind::Constructs
-                } else {
-                    EdgeKind::CallsName
-                };
-                out.push(symbol_edge_with_context(
-                    symbols,
-                    node,
-                    text,
-                    name,
-                    edge_kind,
-                    EdgeConfidence::NameOnly,
-                    EdgeContext {
-                        target_qualified_name: dotted_qualified_name(&identifiers),
-                        receiver_hint: identifiers
-                            .first()
-                            .filter(|_| identifiers.len() > 1)
-                            .cloned(),
-                    },
-                    // The callee is the final segment — `.last()`, matching `identifiers.last()`.
-                    identifier_nodes.last().copied().map(CalleeRange::of_node),
-                ));
-            }
-            if let Some(receiver) = identifiers.first().filter(|_| identifiers.len() > 1).cloned() {
-                out.push(symbol_edge(
-                    symbols,
-                    node,
-                    receiver,
-                    EdgeKind::ReferencesType,
-                    EdgeConfidence::NameOnly,
-                    // The type is the receiver — the FIRST segment, matching
-                    // `identifiers.first()`.
-                    identifier_nodes
-                        .first()
-                        .filter(|_| identifier_nodes.len() > 1)
-                        .copied()
-                        .map(CalleeRange::of_node),
-                ));
-            }
-        },
-        "jsx_opening_element" | "jsx_self_closing_element" => {
-            if let Some(name) = first_identifier_text(node, text) {
-                out.push(symbol_edge(
-                    symbols,
-                    node,
-                    name,
-                    EdgeKind::ReferencesType,
-                    EdgeConfidence::NameOnly,
-                    first_identifier_node(node).map(CalleeRange::of_node),
-                ));
-            }
-        },
-        "type_identifier" => {
-            if let Some(name) = node.utf8_text(text.as_bytes()).ok().map(ToOwned::to_owned) {
-                out.push(symbol_edge(
-                    symbols,
-                    node,
-                    name,
-                    EdgeKind::ReferencesType,
-                    EdgeConfidence::NameOnly,
-                    // `node` is itself the `type_identifier` token — its range is the callee
-                    // range.
-                    Some(CalleeRange::of_node(node)),
-                ));
-            }
-        },
-        _ => {},
-    }
-}
-pub(crate) fn kotlin_edges(
-    text: &str,
-    node: Node<'_>,
-    symbols: &[IndexedSymbol],
-    path: &Path,
-    out: &mut Vec<EdgeCandidate>,
-) {
-    match node.kind() {
-        "import" | "import_header" | "import_directive" => {
-            for name in identifiers_under(node, text) {
-                out.push(file_edge(
-                    path,
-                    node,
-                    text,
-                    name,
-                    EdgeKind::Imports,
-                    EdgeConfidence::NameOnly,
-                ));
-            }
-        },
-        "call_expression" => {
-            let identifiers = identifiers_under(node, text);
-            // Parallel to `identifiers` (same node, same traversal order) so the callee `.last()`
-            // and receiver/constructor `.first()` nodes line up with the string picks (#67).
-            let identifier_nodes = identifier_nodes_under(node);
-            if let Some(name) =
-                identifiers.last().cloned().or_else(|| first_identifier_text(node, text))
-            {
-                out.push(symbol_edge_with_context(
-                    symbols,
-                    node,
-                    text,
-                    name,
-                    EdgeKind::CallsName,
-                    EdgeConfidence::NameOnly,
-                    EdgeContext {
-                        target_qualified_name: dotted_qualified_name(&identifiers),
-                        receiver_hint: identifiers
-                            .first()
-                            .filter(|_| identifiers.len() > 1)
-                            .cloned(),
-                    },
-                    identifier_nodes.last().copied().map(CalleeRange::of_node),
-                ));
-            }
-            if let Some(receiver) = identifiers.first().filter(|_| identifiers.len() > 1).cloned() {
-                out.push(symbol_edge(
-                    symbols,
-                    node,
-                    receiver,
-                    EdgeKind::ReferencesType,
-                    EdgeConfidence::NameOnly,
-                    identifier_nodes
-                        .first()
-                        .filter(|_| identifier_nodes.len() > 1)
-                        .copied()
-                        .map(CalleeRange::of_node),
-                ));
-            }
-            if let Some(constructor) =
-                identifiers.first().filter(|name| looks_like_type_name(name)).cloned()
-            {
-                // Both the type reference and the construct point at the constructor — the FIRST
-                // identifier (matching `identifiers.first()`).
-                let constructor_range = identifier_nodes.first().copied().map(CalleeRange::of_node);
-                out.push(symbol_edge(
-                    symbols,
-                    node,
-                    constructor.clone(),
-                    EdgeKind::ReferencesType,
-                    EdgeConfidence::NameOnly,
-                    constructor_range,
-                ));
-                out.push(symbol_edge_with_context(
-                    symbols,
-                    node,
-                    text,
-                    constructor,
-                    EdgeKind::Constructs,
-                    EdgeConfidence::NameOnly,
-                    EdgeContext::default(),
-                    constructor_range,
-                ));
-            }
-        },
-        "user_type" | "type_identifier" =>
-            if let Some(name) = last_identifier_text(node, text) {
-                out.push(symbol_edge(
-                    symbols,
-                    node,
-                    name,
-                    EdgeKind::ReferencesType,
-                    EdgeConfidence::NameOnly,
-                    last_identifier_node(node).map(final_segment_node).map(CalleeRange::of_node),
-                ));
-            },
-        "delegation_specifier" | "supertype" | "super_type" => {
-            if let Some(name) = last_identifier_text(node, text) {
-                out.push(symbol_edge(
-                    symbols,
-                    node,
-                    name,
-                    EdgeKind::Implements,
-                    EdgeConfidence::NameOnly,
-                    last_identifier_node(node).map(final_segment_node).map(CalleeRange::of_node),
-                ));
-            }
-        },
-        _ => {},
-    }
-}
-pub(crate) fn c_like_edges(
-    text: &str,
-    node: Node<'_>,
-    symbols: &[IndexedSymbol],
-    path: &Path,
-    out: &mut Vec<EdgeCandidate>,
-) {
-    match node.kind() {
-        "preproc_include" => {
-            let include = node_text(node, text)
-                .trim()
-                .trim_start_matches("#include")
-                .trim()
-                .trim_matches(['<', '>', '"'])
-                .to_string();
-            if !include.is_empty() {
-                out.push(file_edge(
-                    path,
-                    node,
-                    text,
-                    include,
-                    EdgeKind::Imports,
-                    EdgeConfidence::NameOnly,
-                ));
-            }
-        },
-        "call_expression" => {
-            let function = node.child_by_field_name("function").unwrap_or(node);
-            let identifiers = identifiers_under(function, text);
-            // Parallel to `identifiers` so the callee `.last()` node matches the string pick (#67).
-            let identifier_nodes = identifier_nodes_under(function);
-            if let Some(name) = identifiers.last().cloned().or_else(|| call_target_name(node, text))
-            {
-                out.push(symbol_edge_with_context(
-                    symbols,
-                    node,
-                    text,
-                    name,
-                    EdgeKind::CallsName,
-                    EdgeConfidence::NameOnly,
-                    EdgeContext {
-                        target_qualified_name: c_like_qualified_name(&identifiers),
-                        receiver_hint: identifiers
-                            .first()
-                            .filter(|_| identifiers.len() > 1)
-                            .cloned(),
-                    },
-                    identifier_nodes.last().copied().map(CalleeRange::of_node),
-                ));
-            }
-        },
-        "type_identifier" | "qualified_identifier" | "namespace_identifier" => {
-            if let Some(name) = last_identifier_text(node, text) {
-                out.push(symbol_edge(
-                    symbols,
-                    node,
-                    name,
-                    EdgeKind::ReferencesType,
-                    EdgeConfidence::NameOnly,
-                    last_identifier_node(node).map(final_segment_node).map(CalleeRange::of_node),
-                ));
-            }
-        },
-        _ => {},
     }
 }
