+++ crates/rag-rat-core/src/index/edges/extract/typescript.rs
@@ -1,0 +1,111 @@
+//! TypeScript graph-edge extraction — the `Language::TypeScript` arm of `syntactic_edges`.
+//! Split out of edges/extract.
+use super::*;
+
+pub(super) fn typescript_edges(
+    text: &str,
+    node: Node<'_>,
+    symbols: &[IndexedSymbol],
+    path: &Path,
+    out: &mut Vec<EdgeCandidate>,
+) {
+    match node.kind() {
+        "import_statement" =>
+            for name in identifiers_under(node, text) {
+                out.push(file_edge(
+                    path,
+                    node,
+                    text,
+                    name,
+                    EdgeKind::Imports,
+                    EdgeConfidence::NameOnly,
+                ));
+            },
+        "export_statement" =>
+            for name in identifiers_under(node, text) {
+                out.push(file_edge(
+                    path,
+                    node,
+                    text,
+                    name,
+                    EdgeKind::Exports,
+                    EdgeConfidence::NameOnly,
+                ));
+            },
+        "call_expression" | "new_expression" => {
+            let function = node.child_by_field_name("function").unwrap_or(node);
+            let identifiers = identifiers_under(function, text);
+            // Parallel to `identifiers` (same traversal order), so `.last()`/`.first()` pick the
+            // node for the same token the string Vec does (#67).
+            let identifier_nodes = identifier_nodes_under(function);
+            if let Some(name) = identifiers.last().cloned().or_else(|| call_target_name(node, text))
+            {
+                let edge_kind = if node.kind() == "new_expression" {
+                    EdgeKind::Constructs
+                } else {
+                    EdgeKind::CallsName
+                };
+                out.push(symbol_edge_with_context(
+                    symbols,
+                    node,
+                    text,
+                    name,
+                    edge_kind,
+                    EdgeConfidence::NameOnly,
+                    EdgeContext {
+                        target_qualified_name: dotted_qualified_name(&identifiers),
+                        receiver_hint: identifiers
+                            .first()
+                            .filter(|_| identifiers.len() > 1)
+                            .cloned(),
+                    },
+                    // The callee is the final segment — `.last()`, matching `identifiers.last()`.
+                    identifier_nodes.last().copied().map(CalleeRange::of_node),
+                ));
+            }
+            if let Some(receiver) = identifiers.first().filter(|_| identifiers.len() > 1).cloned() {
+                out.push(symbol_edge(
+                    symbols,
+                    node,
+                    receiver,
+                    EdgeKind::ReferencesType,
+                    EdgeConfidence::NameOnly,
+                    // The type is the receiver — the FIRST segment, matching
+                    // `identifiers.first()`.
+                    identifier_nodes
+                        .first()
+                        .filter(|_| identifier_nodes.len() > 1)
+                        .copied()
+                        .map(CalleeRange::of_node),
+                ));
+            }
+        },
+        "jsx_opening_element" | "jsx_self_closing_element" => {
+            if let Some(name) = first_identifier_text(node, text) {
+                out.push(symbol_edge(
+                    symbols,
+                    node,
+                    name,
+                    EdgeKind::ReferencesType,
+                    EdgeConfidence::NameOnly,
+                    first_identifier_node(node).map(CalleeRange::of_node),
+                ));
+            }
+        },
+        "type_identifier" => {
+            if let Some(name) = node.utf8_text(text.as_bytes()).ok().map(ToOwned::to_owned) {
+                out.push(symbol_edge(
+                    symbols,
+                    node,
+                    name,
+                    EdgeKind::ReferencesType,
+                    EdgeConfidence::NameOnly,
+                    // `node` is itself the `type_identifier` token — its range is the callee
+                    // range.
+                    Some(CalleeRange::of_node(node)),
+                ));
+            }
+        },
+        _ => {},
+    }
+}
